works with single server
Need to test with multiple servers to verify the working of sharding.
Should start the servers with different ports (for now). Since all the
servers are in the same machine, if all servers start on same port,
client does't know which port to connect.

OR

let master pass the port number when sending server information.